* origin/topic/timw/clang-tidy:
  Don't specify 0u for initial value of loops changed to size_t
  Use .empty() instead of checking size against zero
  Use properly-sized loop variables or convert to ranged-for (bugprone-too-small-loop-variable)

Merge changes one type over to ptrdiff_t

+++ src/OpaqueVal.cc
@@ -965,7 +965,7 @@ VectorValPtr ParaglobVal::Get(StringVal* &pattern)
 	std::string string_pattern (reinterpret_cast<const char*>(pattern->Bytes()), pattern->Len());
 
 	std::vector<std::string> matches = this->internal_paraglob->get(string_pattern);
-	for (unsigned int i = 0; i < matches.size(); i++)
+	for ( size_t i = 0; i < matches.size(); i++ )
 		rval->Assign(i, zeek::make_intrusive<StringVal>(matches.at(i)));
 
 	return rval;
